requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
nanoid==2.0.0
langchain==0.0.170
openai==0.27.6
pandas==2.0.1
tabulate==0.9.0
python-dotenv==1.0.0